la_fire_monitoring/requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
scikit-learn
torch
torchvision
tqdm
rasterio
geopandas
folium
pyproj
shapely
opencv-python
scipy
pillow
tensorflow  # for LSTM forecasting (optional, you may switch to PyTorch)
notebook
